fix: Switch to more appropriate data types

+++ src/format.cpp
@@ -6,8 +6,8 @@
 
 using std::string;
 
-string Format::ElapsedTime(long seconds) {
-  const unsigned int elapsedHours = seconds / 3600;
+string Format::ElapsedTime(unsigned long long int seconds) {
+  const unsigned long long int elapsedHours = seconds / 3600;
   const unsigned short elapsedMinutes = (seconds % 3600) / 60;
   const unsigned short elapsedSeconds = (seconds % 3600) % 60;
